+++ dbentry/autocomplete/widgets.py
@@ -22,14 +22,14 @@ DEFAULTS = {
 
 
 def make_widget(
-    model,
-    tabular=False,
-    multiple=False,
-    namespace="admin",
-    can_add=True,
-    can_list=True,
-    can_edit=True,
-    **kwargs,
+        model,
+        tabular=False,
+        multiple=False,
+        namespace="admin",
+        can_add=True,
+        can_list=True,
+        can_edit=True,
+        **kwargs,
 ):
     """
     Factory function that creates MIZSelect autocomplete widgets.