+++ dbentry/tools/views.py
@@ -409,8 +409,8 @@ class SiteSearchView(views.generic.TemplateView):
             codename_view = get_permission_codename('view', opts)
             codename_change = get_permission_codename('change', opts)
             return (
-                user.has_perm('%s.%s' % (opts.app_label, codename_view))
-                or user.has_perm('%s.%s' % (opts.app_label, codename_change))
+                    user.has_perm('%s.%s' % (opts.app_label, codename_view))
+                    or user.has_perm('%s.%s' % (opts.app_label, codename_change))
             )
 
         app = apps.get_app_config(app_label or self.app_label)