+++ tests/test_actions/test_actions.py
@@ -12,6 +12,7 @@ def func0():
 
 def func1():
     pass
+# TODO: use a decorator to add attributes to the function
 func1.short_description = 'Foo'  # noqa
 func1.allowed_permissions = 'Bar'