+++ tests/test_dbentry/test_query.py
@@ -1,7 +1,7 @@
 import datetime
 import random
 from itertools import chain
-from unittest.mock import patch, Mock
+from unittest.mock import patch
 
 from django.core.exceptions import FieldDoesNotExist, FieldError
 from django.db import models